requirements.txt
Flask==2.1.3
babel==2.9.0
python-dateutil==2.6.0
flask-moment==0.11.0
flask-wtf==0.14.3
flask_sqlalchemy==2.4.4
SQLAlchemy==1.4.39
postgres==4.0
Flask-Migrate==3.1.0